refactor: Api Response 변수 추가, 공통반환 설정
