--- e-commerce/src/main/java/backend/core/common/errorcode/ErrorResponse.java
@@ -1,39 +1,0 @@
-package backend.core.common.errorcode;
-
-import com.fasterxml.jackson.annotation.JsonInclude;
-import java.util.List;
-import lombok.Builder;
-import lombok.Getter;
-import lombok.RequiredArgsConstructor;
-import org.springframework.validation.FieldError;
-
-@Getter
-@Builder
-@RequiredArgsConstructor
-public class ErrorResponse {
-
-    private final String code;
-    private final String message;
-
-    @JsonInclude(JsonInclude.Include.NON_EMPTY)
-    private final List<ValidationError> errors;
-
-    @JsonInclude(JsonInclude.Include.NON_EMPTY)
-    private final Object details; // 추가 데이터 (userId 등)
-
-    @Getter
-    @Builder
-    @RequiredArgsConstructor
-    public static class ValidationError {
-
-        private final String field;
-        private final String message;
-
-        public static ValidationError of(final FieldError fieldError) {
-            return ValidationError.builder()
-                    .field(fieldError.getField())
-                    .message(fieldError.getDefaultMessage())
-                    .build();
-        }
-    }
-}
